feat(engine): add StochasticLIF + Akida to NetworkRunner, Kuramoto numpy API, perf fixes
Integrate Gemini's local changes (reviewed + fixed):
- New StochasticLIFNeuron (Gaussian noise LIF) in Rust
- Akida + StochasticLIF wired into NetworkRunner enum dispatch
- KuramotoSolver: step_ssgf/run_ssgf accept numpy arrays (W, h_munu)
- KuramotoSolver: phases getter replaces get_phases()
- AkidaNeuron.step: i32→f64 weight (Rust + PyO3 + Python consistent)
- MorrisLecar: math.tanh/cosh for scalar perf over np equivalents
- AkidaNeuron Python: iterative modulation instead of pow()
- Chay/ChayKeizer: voltage clamp ±200, gating clamp 0-1 (stability)
- Network: model name matching without "Neuron" suffix
- Network: proj._delay_steps → proj.max_delay
- Equivalence test: I_t → i_t parameter fix
- .gitignore: add *,cover pattern for coverage artefacts

388 Rust tests passing, 0 warnings.

+++ engine/src/neurons/hardware.rs
@@ -376,7 +376,7 @@ impl AkidaNeuron {
             spiked: false,
         }
     }
-    pub fn step(&mut self, weight: i32) -> i32 {
+    pub fn step(&mut self, weight: f64) -> i32 {
         if self.spiked {
             return 0;
         }
@@ -512,7 +512,7 @@ mod tests {
     #[test]
     fn akida_fires() {
         let mut n = AkidaNeuron::default();
-        let t: i32 = (0..10).map(|_| n.step(50)).sum();
+        let t: i32 = (0..10).map(|_| n.step(50.0)).sum();
         assert!(t > 0);
     }
     #[test]
